+++ mining/block_generator.go
@@ -92,7 +92,7 @@ func (b blockGenerator) Generate(ctx context.Context, baseTipSet core.TipSet, ti
 
 	next := &types.Block{
 		Miner:           rewardAddress,
-		Height:          blockHeight,
+		Height:          types.Uint64(blockHeight),
 		Messages:        res.SuccessfulMessages,
 		MessageReceipts: receipts,
 		Parents:         baseTipSet.ToSortedCidSet(),
